python/requirements.txt
azure-servicebus
azure-eventhub-checkpointstoreblob
django
opentelemetry-sdk
opentelemetry-instrumentation-django
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-urllib3
azure-core-tracing-opentelemetry
azure-monitor-opentelemetry-exporter
jaeger
